+++ lang/zh_CN/Audio/ExpansionPackageLines.lua
@@ -307,10 +307,4 @@ return {
     ['$LuaGuanzong1'] = '汝为叔父，怎可与小辈计较！',
     ['$LuaGuanzong2'] = '阿瞒生龙活虎，汝切勿胡言！',
     ['~ExCaosong'] = '长恨人心不如水，等闲平地起波澜……',
-    -- 蒲元
-    ['$LuaZhuren1'] = '造刀三千口，用法各不同',
-    ['$LuaZhuren2'] = '此刀，可劈铁珠之筒！',
-    ['$LuaTianjiang1'] = '天赐匠法，精心锤炼！',
-    ['$LuaTianjiang2'] = '巧夺天工，超凡脱俗！',
-    ['~ExPuyuan'] = '铸木镂冰，怎成大器！',
 }
