+++ lang/zh_CN/Audio/ExtraordinaryPeoplePackageLines.lua
@@ -1,0 +1,10 @@
+-- lines for Extraordinary People Package
+
+return {
+    -- 蒲元
+    ['$LuaZhuren1'] = '造刀三千口，用法各不同',
+    ['$LuaZhuren2'] = '此刀，可劈铁珠之筒！',
+    ['$LuaTianjiang1'] = '天赐匠法，精心锤炼！',
+    ['$LuaTianjiang2'] = '巧夺天工，超凡脱俗！',
+    ['~ExPuyuan'] = '铸木镂冰，怎成大器！',
+}
