+++ lang/zh_CN/Package/ExpansionPackage.lua
@@ -628,14 +628,4 @@ return {
     ['LuaGuanzong'] = '惯纵',
     [':LuaGuanzong'] = '出牌阶段限一次，你可以令一名其他角色视为对另一名你选择的其他角色造成过1点伤害',
     ['#LuaGuanzong'] = '%from 执行“%arg”的效果，视为对 %to 造成了 %arg2 点伤害',
-    ['ExPuyuan'] = '蒲元',
-    ['&ExPuyuan'] = '蒲元',
-    ['#ExPuyuan'] = '淬炼百兵',
-    ['LuaTianjiang'] = '天匠',
-    [':LuaTianjiang'] = '①游戏开始时，你随机获得两张不同副类别的装备牌并置入你的装备区；\
-    ②出牌阶段，你可以将装备区里的牌移至其他角色的装备区（可以替换原装备），然后若此牌是“铸刃”的武器牌，你摸两张牌',
-    ['luatianjiang'] = '天匠',
-    ['LuaZhuren'] = '铸刃',
-    [':LuaZhuren'] = '出牌阶段限一次，你可以弃置一张手牌，打造一张武器牌（若打造失败或武器已有则改为获得一张【杀】，花色决定武器名称，点数决定成功率；当“铸刃”的武器牌进入弃牌堆时，将之移出游戏）',
-    ['luazhuren'] = '铸刃',
 }
