+++ lang/zh_CN/Package/ExtraordinaryPeoplePackage.lua
@@ -1,0 +1,15 @@
+-- translation for ExtraordinaryPeoplePackage
+
+return {
+    ['ExtraordinaryPeoplePackage'] = '限定-奇人异士包',
+    ['ExPuyuan'] = '蒲元',
+    ['&ExPuyuan'] = '蒲元',
+    ['#ExPuyuan'] = '淬炼百兵',
+    ['LuaTianjiang'] = '天匠',
+    [':LuaTianjiang'] = '①游戏开始时，你随机获得两张不同副类别的装备牌并置入你的装备区；\
+    ②出牌阶段，你可以将装备区里的牌移至其他角色的装备区（可以替换原装备），然后若此牌是“铸刃”的武器牌，你摸两张牌',
+    ['luatianjiang'] = '天匠',
+    ['LuaZhuren'] = '铸刃',
+    [':LuaZhuren'] = '出牌阶段限一次，你可以弃置一张手牌，打造一张武器牌（若打造失败或武器已有则改为获得一张【杀】，花色决定武器名称，点数决定成功率；当“铸刃”的武器牌进入弃牌堆时，将之移出游戏）',
+    ['luazhuren'] = '铸刃',
+}
